# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Retriever.sh",
    "description": "Cheap, easy to use hybrid search engine",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/rag/projects/{project_id}/items": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Ingest Item",
        "operationId": "ingest_item_api_rag_projects__project_id__items_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ItemIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rag/projects/{project_id}/items/{item_id}": {
      "delete": {
        "tags": [
          "rag"
        ],
        "summary": "Delete Item",
        "operationId": "delete_item_api_rag_projects__project_id__items__item_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Item Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rag/projects/{project_id}/auth/check": {
      "get": {
        "tags": [
          "rag"
        ],
        "summary": "Check Project Api Key",
        "operationId": "check_project_api_key_api_rag_projects__project_id__auth_check_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rag/projects/{project_id}/query": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Query Project",
        "operationId": "query_project_api_rag_projects__project_id__query_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/QueryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioBase64ContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "audio_base64",
            "title": "Type"
          },
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          },
          "media_type": {
            "type": "string",
            "minLength": 1,
            "title": "Media Type"
          }
        },
        "type": "object",
        "required": [
          "type",
          "data",
          "media_type"
        ],
        "title": "AudioBase64ContentBlock"
      },
      "AudioUrlContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "audio_url",
            "title": "Type"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "AudioUrlContentBlock"
      },
...
      "FileBase64ContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "file_base64",
            "title": "Type"
          },
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          },
          "media_type": {
            "type": "string",
            "const": "application/pdf",
            "title": "Media Type",
            "default": "application/pdf"
          }
        },
        "type": "object",
        "required": [
          "type",
          "data"
        ],
        "title": "FileBase64ContentBlock"
      },
      "FileUrlContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "file_url",
            "title": "Type"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url"
          },
          "media_type": {
            "type": "string",
            "const": "application/pdf",
            "title": "Media Type",
            "default": "application/pdf"
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "FileUrlContentBlock"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageBase64ContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "image_base64",
            "title": "Type"
          },
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          },
          "media_type": {
            "type": "string",
            "minLength": 1,
            "title": "Media Type"
          }
        },
        "type": "object",
        "required": [
          "type",
          "data",
          "media_type"
        ],
        "title": "ImageBase64ContentBlock"
      },
      "ImageUrlContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "image_url",
            "title": "Type"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "ImageUrlContentBlock"
      },
      "ItemIn": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileBase64ContentBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "audio_base64": "#/components/schemas/AudioBase64ContentBlock",
                  "audio_url": "#/components/schemas/AudioUrlContentBlock",
                  "file_base64": "#/components/schemas/FileBase64ContentBlock",
                  "file_url": "#/components/schemas/FileUrlContentBlock",
                  "image_base64": "#/components/schemas/ImageBase64ContentBlock",
                  "image_url": "#/components/schemas/ImageUrlContentBlock",
                  "text": "#/components/schemas/TextContentBlock",
                  "video_base64": "#/components/schemas/VideoBase64ContentBlock",
                  "video_url": "#/components/schemas/VideoUrlContentBlock"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Content"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "ItemIn"
      },
      "ItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileBase64ContentBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "audio_base64": "#/components/schemas/AudioBase64ContentBlock",
                  "audio_url": "#/components/schemas/AudioUrlContentBlock",
                  "file_base64": "#/components/schemas/FileBase64ContentBlock",
                  "file_url": "#/components/schemas/FileUrlContentBlock",
                  "image_base64": "#/components/schemas/ImageBase64ContentBlock",
                  "image_url": "#/components/schemas/ImageUrlContentBlock",
                  "text": "#/components/schemas/TextContentBlock",
                  "video_base64": "#/components/schemas/VideoBase64ContentBlock",
                  "video_url": "#/components/schemas/VideoUrlContentBlock"
                }
              }
            },
            "type": "array",
            "title": "Content"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "metadata",
          "created_at"
        ],
        "title": "ItemOut"
      },
...
      "QueryRequest": {
        "properties": {
          "input": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/AudioBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/VideoBase64ContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileUrlContentBlock"
                },
                {
                  "$ref": "#/components/schemas/FileBase64ContentBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "audio_base64": "#/components/schemas/AudioBase64ContentBlock",
                  "audio_url": "#/components/schemas/AudioUrlContentBlock",
                  "file_base64": "#/components/schemas/FileBase64ContentBlock",
                  "file_url": "#/components/schemas/FileUrlContentBlock",
                  "image_base64": "#/components/schemas/ImageBase64ContentBlock",
                  "image_url": "#/components/schemas/ImageUrlContentBlock",
                  "text": "#/components/schemas/TextContentBlock",
                  "video_base64": "#/components/schemas/VideoBase64ContentBlock",
                  "video_url": "#/components/schemas/VideoUrlContentBlock"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Input"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "vector_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector K"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/QueryResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "QueryResponse"
      },
...
      "TextContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "text",
            "title": "Type"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "type",
          "text"
        ],
        "title": "TextContentBlock"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoBase64ContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "video_base64",
            "title": "Type"
          },
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          },
          "media_type": {
            "type": "string",
            "minLength": 1,
            "title": "Media Type"
          }
        },
        "type": "object",
        "required": [
          "type",
          "data",
          "media_type"
        ],
        "title": "VideoBase64ContentBlock"
      },
      "VideoUrlContentBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "video_url",
            "title": "Type"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "VideoUrlContentBlock"
      }
    }
  }
}